requirements.txt
torch>=1.7.0
torchvision>=0.8.0
ultralytics>=8.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pillow>=8.3.0
numpy>=1.19.0
opencv-python>=4.5.0
requests>=2.25.0
google-generativeai>=0.3.0
cloudinary>=1.44.1
tensorflow>=2.15.0
openai>=1.0.0
python-dotenv>=1.0.0
azure-ai-inference>=1.0.0